move all queries to /queries
