+++ src/Layout/Layout.js
@@ -1,8 +1,12 @@
 import React from "react";
-import { gql, useQuery, ApolloConsumer } from "@apollo/client";
+import { useQuery, ApolloConsumer } from "@apollo/client";
 import { connect } from "react-redux";
 import PropTypes from "prop-types";
 import { device } from "../utils/devices";
+import {
+  GET_FEATURED_PRODUCTS_AND_ARTICLES,
+  GET_CHECKOUT,
+} from "../queries";
 
 import * as CartActionCreators from "../state/actions/cart";
 import styled from "styled-components";
@@ -21,71 +25,6 @@ const MasterWrapper = styled.div`
   /* Media screen keeps WelcomeSection's img 100% height on bigger screens */
   @media ${device.laptop} {
     max-width: 100vw;
-  }
-`;
-
-const GET_FEATURED_PRODUCTS_AND_ARTICLES = gql`
-  query GetProductsAndArticles {
-    collections(
-      query: "title:'Wholesale Products' OR title:'Featured Products'"
-      first: 2
-    ) {
-      edges {
-        node {
-          title
-          products(first: 5) {
-            edges {
-              node {
-                id
-                title
-                handle
-                availableForSale
-                variants(first: 1) {
-                  edges {
-                    node {
-                      price
-                    }
-                  }
-                }
-                images(first: 6) {
-                  edges {
-                    node {
-                      transformedSrc(maxWidth: 400, maxHeight: 450)
-                      altText
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    articles(first: 20) {
-      edges {
-        node {
-          title
-          contentHtml
-          excerpt
-          handle
-          image {
-            transformedSrc(maxWidth: 750, maxHeight: 750)
-            altText
-          }
-        }
-      }
-    }
-  }
-`;
-
-const GET_CHECKOUT = gql`
-  query getCheckout($id: ID!) {
-    node(id: $id) {
-      ... on Checkout {
-        completedAt
-        id
-      }
-    }
   }
 `;
