+++ src/pages/Shop/Shop.js
@@ -1,11 +1,12 @@
 import React from "react";
 import { useLocation } from "react-router-dom";
 import styled from "styled-components";
-import { gql, useQuery } from "@apollo/client";
+import { useQuery } from "@apollo/client";
 import PageWrapper from "../../SharedComponents/PageWrapper";
 import ShopProduct from "./ShopProduct";
 import StyledH1 from "../../SharedComponents/StyledH1";
 import Footer from "../../SharedComponents/Footer";
+import { GET_SHOP_PRODUCTS } from "../../queries";
 
 const ProductsContainer = styled.div`
   display: flex;
@@ -17,42 +18,6 @@ const ProductsContainer = styled.div`
 `;
 
 const Shop = () => {
-  const GET_SHOP_PRODUCTS = gql`
-    query getShopProducts($collectionName: String!) {
-      collections(query: $collectionName, first: 2) {
-        edges {
-          node {
-            title
-            products(first: 20) {
-              edges {
-                node {
-                  id
-                  title
-                  handle
-                  availableForSale
-                  variants(first: 1) {
-                    edges {
-                      node {
-                        price
-                      }
-                    }
-                  }
-                  images(first: 6) {
-                    edges {
-                      node {
-                        altText
-                        transformedSrc(maxWidth: 400, maxHeight: 450)
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  `;
 
   const location = useLocation();
