+++ src/state/app.js
@@ -5,13 +5,6 @@ export const initialState = {
     lineItems: [],
     checkoutId: "",
   },
-  onlineStore: [],
-  wholesaleProducts: [],
-  featuredProducts: [],
-  articles: [],
-  pending: true,
-  error: null,
-  lastShopifyFetchTimestamp: 0,
   burgerToggled: false,
   burgerClickedOnce: false,
   heroImgSrc: "",
